requirements.txt
absl-py==1.4.0
altair==5.0.1
ansi2html==1.8.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
catboost==1.2
category-encoders==2.6.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
executing==1.2.0
Faker==19.3.0
fastjsonschema==2.18.0
favicon==0.7.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.57.0
h5py==3.9.0
htbuilder==0.6.1
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.12.2
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaleido==0.2.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lightgbm==3.3.5
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
nbformat==5.9.2
nest-asyncio==1.5.7
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
orjson==3.9.4
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.15.0
plotly-resampler==0.9.1
pmdarima==2.0.3
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycaret==3.0.4
pydeck==0.8.0
Pygments==2.16.1
pymdown-extensions==10.1
Pympler==1.0.1
pyod==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
schemdraw==0.15
scikit-base==0.5.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
six==1.16.0
sktime==0.21.0
smmap==5.0.0
soupsieve==2.4.1
st-annotated-text==4.0.0
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tbats==1.1.3
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.1
trace-updater==0.0.9.1
traitlets==5.9.0
tsdownsample==0.1.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wrapt==1.15.0
xxhash==3.3.0
yellowbrick==1.5
zipp==3.16.2
